+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 
+- Implemented check for bad chronological order
+
 
 ## [v0.2.0] Blame (2023-07-09)